Sorting parameters hierarchy (chain) is now used.
Sorting parameters now allows to create hierarchical structure of
sorting parameters. When two elements cannot be sorted using the sorting
key of the root level SortingParametersChain object they are sorted
based on the sorting key of the child SortingParametersChain object.
This is done recursively. Now it is possible to first sort based on X
component and for elements with the same values for X- to sort them by Y
component.