Bucket sort fix
Sometimes there is an error in the bucket sort final result. The error
seems to be associated with the floating point comparison uncertainty,
because it is only encountered with the elements that are very close to
each other in terms of their sorting key values. It should be further
investigated.